Move `onnx_inference()` into `deepseg_.onnx.py`
