+++ spinalcordtoolbox/deepseg_/sc.py
@@ -12,7 +12,7 @@ from scipy.ndimage.measurements import center_of_mass, label
 from scipy.ndimage import distance_transform_edt
 
 from spinalcordtoolbox import resampling
-from spinalcordtoolbox.deepseg.models import onnx_inference
+from spinalcordtoolbox.deepseg_.onnx import onnx_inference
 from spinalcordtoolbox.deepseg_.postprocessing import post_processing_volume_wise, keep_largest_object, fill_holes_2d
 from spinalcordtoolbox.image import Image, empty_like, change_type, zeros_like, add_suffix, concat_data, split_img_data
 from spinalcordtoolbox.centerline.core import ParamCenterline, get_centerline, _call_viewer_centerline
